update tests for election data now being in a model
Change tests that override settings.ELECTION things to instead mock
the Elections model appropriately

We also need to import the election specific static data at the point
it's used. We need to do it this way because it pulls it from the database
which might not be available when the import happens, e.g during initial
migration and test set up. So, instead of importing the election specific
data at the top of modules do it just before it's used in the function.