rename MAPIT_DATA to AREA_DATA
because not everything uses MapIt and also we are using area type for
the model rename MAPIT_DATA to AREA_DATA so it's less confusing. Also
it's less confusing for reading the code if you don't know what MapIt
is.

The BaseMapitData model remains though as that's the code that calls
MapIt so it makes sense to leave that referring to MapIt